wsd: replace Util::stoi() with Util::i32FromString()

+++ common/JailUtil.cpp
@@ -20,6 +20,7 @@
 
 #include <common/FileUtil.hpp>
 #include <common/Log.hpp>
+#include <common/Util.hpp>
 
 #include <SigUtil.hpp>
 #include <csignal>
@@ -385,19 +386,23 @@ void cleanupJails(const std::string& root)
             if (pidSepPos != std::string::npos)
             {
                 bool skip = false;
-                std::string pidStr = jail.substr(0, pidSepPos);
-                try {
-                    int pid = std::stoi(pidStr);
+                const std::string_view pidStr = std::string_view(jail).substr(0, pidSepPos);
+                const auto [pid, success] = Util::i32FromString(pidStr);
+                if (success && pid > 1)
+                {
                     LOG_TRC("Checking pid for jail " << pid << " " << root);
                     if (pid != getpid() && ::kill(pid, 0) == 0)
                     {
                         LOG_TRC("Skipping cleaning jails directory for running coolwsd with pid " << pid);
                         skip = true;
                     }
-                } catch(...) {
+                }
+                else
+                {
                     // Problematic - may delete a jail that is not ours then ...
                     LOG_WRN("Exception parsing pid '" << pidStr << "' from '" << jail << "'");
                 }
+
                 if (!skip)
                 {
                     std::vector<std::string> newJails;
